requirements.txt
azure-cli
resource-graph
pandas
pyecore
neomodel
requests
requests_ntlm
dash
dash-html-components
dash-cytoscape
xmltodict
python-dotenv